requirements.txt
Flask
python-dotenv
requests
PyMuPDF
docx2txt